patient-service/app.py
import os
from flask import Flask
from flask_jwt_extended import JWTManager
from dotenv import load_dotenv
from models import db
from routes import patient_bp

load_dotenv()

def create_app():
    app = Flask(__name__)
    app.config["SQLALCHEMY_DATABASE_URI"] = os.getenv("DB_URL")
    app.config["JWT_SECRET_KEY"] = os.getenv("JWT_SECRET_KEY")
    app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False

    db.init_app(app)
    JWTManager(app)
    app.register_blueprint(patient_bp, url_prefix="/patient")

    # create tables if needed
    with app.app_context():
        db.create_all()

    return app

app = create_app()

if __name__ == "__main__":
    app.run(host="0.0.0.0", port=5003, debug=True)
